Delivery Quest System v0.5
Periodically picks a random inactive delivery point to assign a quest to and generates an amount of fuel to deliver, a point reward, and a time limit for that quest. Completing the quest works the same way as before (deliver fuel, get points, delivery point becomes inactive again). Failing to deliver the quest within the time limit currently results in an instant game over (will probably change this later to something like "three strikes and you're out").
There's still a lot of UI stuff to do related to this system; will take care of this later.